+++ packages/kn-plugin-workflow/e2e-tests/quarkus_run_test.go
@@ -30,9 +30,9 @@ import (
 	"testing"
 	"time"
 
-	"github.com/kiegroup/kie-tools/packages/kn-plugin-workflow/pkg/command"
-	"github.com/kiegroup/kie-tools/packages/kn-plugin-workflow/pkg/command/quarkus"
-	"github.com/kiegroup/kie-tools/packages/kn-plugin-workflow/pkg/common"
+	"github.com/apache/incubator-kie-tools/packages/kn-plugin-workflow/pkg/command"
+	"github.com/apache/incubator-kie-tools/packages/kn-plugin-workflow/pkg/command/quarkus"
+	"github.com/apache/incubator-kie-tools/packages/kn-plugin-workflow/pkg/common"
 	"github.com/stretchr/testify/require"
 )