+++ packages/runtime-tools-gateway-api/src/__tests__/apis/apis.test.tsx
@@ -88,6 +88,9 @@ describe("swf custom form tests", () => {
 
   it("get custom workflow schema - success - with workflowdata", async () => {
     const schema = {
+      components: {
+        schemas: {},
+      },
       type: "object",
       properties: {
         name: {